groupchat_app_src/backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy>=2.0.0
asyncmy
asyncpg
greenlet
bcrypt==4.0.1
passlib
python-jose[cryptography]
python-dotenv
httpx
jinja2
